+++ frontend/lib/models/purchased_item.dart
@@ -13,7 +13,7 @@ class PurchasedItem {
   final int trip;
   final int pantryItem;
   final double purchasePrice;
-  final int quantityBought;
+  final double quantityBought;
   final int? createdBy;
   final int? updatedBy;
 }
